refactor: update rexexp match for jira references (#3)
- Fix regular expression to match entire Jira ticket link in the pull
request body.
- Added tests for invalid Jira ticket references in pull request body.
- Fix identified security issue: `Prototype Pollution in JSON5 via Parse
Method`:
https://github.com/finderau/github-pr-linter/security/dependabot/1

**Security issue summary:**
The parse method of the JSON5 library before and including version 2.2.1
does not restrict parsing of keys named __proto__, allowing specially
crafted strings to pollute the prototype of the resulting object.

This vulnerability pollutes the prototype of the object returned by
JSON5.parse and not the global Object prototype, which is the commonly
understood definition of Prototype Pollution. However, polluting the
prototype of a single object can have significant security impact for an
application if the object is later used in trusted operations.